Move wallet components into wallets/
