+++ pages/wallets/[wallet].js
@@ -2,7 +2,7 @@ import { getGetServerSideProps } from '@/api/ssrApollo'
 import { Form, ClientInput, PasswordInput, CheckboxGroup, Checkbox } from '@/components/form'
 import { CenterLayout } from '@/components/layout'
 import { WalletSecurityBanner } from '@/components/banners'
-import { WalletLogs } from '@/components/wallet-logger'
+import { WalletLogs } from '@/wallets/logger'
 import { useToast } from '@/components/toast'
 import { useRouter } from 'next/router'
 import { useWallet } from '@/wallets/index'
@@ -11,14 +11,14 @@ import Text from '@/components/text'
 import { autowithdrawInitial, AutowithdrawSettings } from '@/components/autowithdraw-shared'
 import { canReceive, canSend, isConfigured } from '@/wallets/common'
 import { SSR } from '@/lib/constants'
-import WalletButtonBar from '@/components/wallet-buttonbar'
+import WalletButtonBar from '@/wallets/buttonbar'
 import { useWalletConfigurator } from '@/wallets/config'
 import { useCallback, useMemo } from 'react'
 import { useMe } from '@/components/me'
 import validateWallet from '@/wallets/validate'
 import { ValidationError } from 'yup'
 import { useFormikContext } from 'formik'
-import { useWalletImage } from '@/components/wallet-image'
+import { useWalletImage } from '@/wallets/image'
 import styles from '@/styles/wallet.module.css'
 
 export const getServerSideProps = getGetServerSideProps({ authRequired: true })
